♻️ [Refactoring]: #306 toastSink を root barrel 経由の import に統一
lib/tauri 公開 API の規約「src/lib/tauri/index.ts を唯一の公開 API とする」
から外れていた toastSink の import 経路を是正する。

- root barrel に registerToastSink / unregisterToastSink / ToastSink 型を export
- App.tsx と App 統合テスト 2 件の import を @/lib/tauri/toastSink から
  @/lib/tauri へ変更（外部消費者を barrel 経由に統一）
- index.ts の孤立した // dialog セクションコメントを openDirectoryDialog
  export の直前へ移動し、コメントと export の対応を整理

getToastSink は invokeWrapped が循環参照回避のため個別 alias で内部参照する
内部専用 API のため barrel には公開しない（CLAUDE.md の既存の例外規定どおり）。
外部挙動は不変。
