+++ src/__tests__/App.mutationFailureToast.test.tsx
@@ -11,7 +11,7 @@ import {
 } from "vitest";
 import { App } from "@/App";
 import { DRAG_MIME_TYPE } from "@/features/board/components/Board/dragState";
-import { unregisterToastSink } from "@/lib/tauri/toastSink";
+import { unregisterToastSink } from "@/lib/tauri";
 import { createDragEvent } from "@/test-fixtures/createDragEvent";
 import type { TaskPayload } from "@/types/task";
